+++ src/components/Footer.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export default function Footer() {
+	return (
+		<footer>
+			<div id="footer-wrapper">
+				<div id="contact-form-wrapper">contact form area</div>
+				<div id="social-contact-wrapper">social contact</div>
+				<div id="copyright-wrapper">copyright area</div>
+			</div>
+		</footer>
+	)
+}