+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## March 29, 2020
+* 🌱 Cread Card, Footer, Header components
+* 🐩 Working on Card styling
+* 🧂 Added react-google-font-loader
+
 ## March 27, 2020
 * 🌱 Created Router, Home, NotFound and Photography components